Commented and finished first build
I commented out all the broken code for the gui, as I can't get that
working, period. Made a simple bypass for now so it only opens the link
in the browser. Not the most elegant solution, but it works for now.
Continuing developing of the gui on another branch another time.